# secure-gate-access/api-documentation.yaml
openapi: 3.0.3
info:
  title: Secure Gate Access API
  description: |
    Comprehensive API for visitor management system with role-based access control.
    
    ## Authentication
    This API uses JWT Bearer tokens for authentication. Include the token in the Authorization header:
    `Authorization: Bearer <your_jwt_token>`
    
    ## Roles
    - **resident**: Can create visitors and manage their invitations
    - **guard**: Can check visitors in/out and view active visitors  
    - **admin**: Full system access including metrics and audit logs
    
  version: 3.0.0
  contact:
    name: Secure Gate Access Team
    email: support@securegate.com
  license:
    name: MIT
servers:
  - url: http://localhost:5000/api
    description: Development server
  - url: https://api.securegate.com/api
    description: Production server
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Visitor:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        email:
          type: string
          format: email
        id_number:
          type: string
        vehicle_plate:
          type: string
        status:
          type: string
          enum: [PENDING, CONFIRMED, ON_PREMISE, EXITED, REVOKED, EXPIRED]
        host_email:
          type: string
        date_of_visit:
          type: string
          format: date
        time:
          type: string
        purpose:
          type: string
        otp:
          type: string
        otp_expires_at:
          type: string
          format: date-time
        check_in_time:
          type: string
          format: date-time
        check_out_time:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
paths:
...
  /visitors:
    post:
      tags:
        - Visitor Management
      summary: Create visitor invitation
      description: Create a new visitor invitation (Resident only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                phone:
                  type: string
                email:
                  type: string
                  format: email
                id_number:
                  type: string
                vehicle_plate:
                  type: string
                date_of_visit:
                  type: string
                  format: date
                time:
                  type: string
                purpose:
                  type: string
              required: [name, phone, date_of_visit, time, purpose]
      responses:
        '201':
          description: Visitor invitation created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Visitor'

    get:
      tags:
        - Visitor Management
      summary: Get my visitors
      description: Get all visitors created by the authenticated resident
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: status
          in: query
          schema:
            type: string
            enum: [PENDING, CONFIRMED, ON_PREMISE, EXITED, REVOKED, EXPIRED]
      responses:
        '200':
          description: Visitors retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Visitor'
...